Implemented 'bldsup' local plugin support.
