+++ build.py
@@ -14,11 +14,7 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-import os
-import subprocess
-
-from pybuilder.core import init, use_plugin, Author, task
-from pybuilder.utils import assert_can_execute
+from pybuilder.core import init, use_bldsup, use_plugin, Author
 
 use_plugin("python.core")
 use_plugin("python.pytddmon")
@@ -39,6 +35,8 @@ use_plugin("python.pydev")
 use_plugin("python.pycharm")
 use_plugin("python.pytddmon")
 
+use_bldsup()
+use_plugin("pdoc")
 
 summary = "An extensible, easy to use continuous build tool for Python"
 description = """PyBuilder is a continuous build tool for multiple languages.
@@ -109,19 +107,3 @@ def initialize(project):
                          'Topic :: Software Development :: Build Tools',
                          'Topic :: Software Development :: Quality Assurance',
                          'Topic :: Software Development :: Testing'])
-
-
-@task
-def pdoc_generate(project, logger):
-    assert_can_execute(command_and_arguments=["pdoc", "--version"],
-                       prerequisite="pdoc",
-                       caller=pdoc_generate.__name__)
-
-    logger.info("Generating pdoc documentation")
-
-    command_and_arguments = ["pdoc", "--html", "pybuilder", "--all-submodules", "--overwrite", "--html-dir", "api-doc"]
-    source_directory = project.get_property("dir_source_main_python")
-    environment = {"PYTHONPATH": source_directory,
-                   "PATH": os.environ["PATH"]}
-
-    subprocess.check_call(command_and_arguments, shell=False, env=environment)
